requirements.txt
Flask==3.0.3
requests==2.28.1
python-docx==1.1.2
beautifulsoup4==4.12.2